bot/requirements.txt
python-telegram-bot==20.7
gspread==6.1.2
google-auth==2.29.0
apscheduler==3.10.4
python-dotenv==1.0.1
